[IRGenSIL] Ignore ptr casts when generating shadow copies
We decide whether or not to generate shadow copies based on whether an
instruction is an Alloca or not. We should ignore all pointer casts when doing
so, otherwise we may artificially reach the conclusion that the instruction does
not represent an alloca.

Furthermore, once we reach the alloca, we should also check if it is a static
alloca: in that case, no shadow copy is needed.

It may look like this change is not tested, but it is. We were forgetting to do
two things (strip ptr casts + static alloca check). Fixing only one of those
would introduce test failures. Fixing both maintain the existing test behavior.