[IRGenDebugInfo] Look through ptr casts
When deciding whether to generate dbg.declare or dbg.values, we check whether
the variable "storage" is an alloca or not. But this fails to account for no-op
pointer casts.

Some of the tests I updated were expecting really weird behavior. For example,
debug_info_expression was generating an alloca + dbg declare, followed by a
debug value for the same variable.