AST: Inherit access level of opaque type decls from naming decl.
When an `OpaqueTypeDecl` is constructed, the access level attributes of the
decl that names the opaque type are copied on to it. However, the
`@usableFromInline` attribute is not permitted on every decl, so it does not get
copied. This in turn causes effective access level computations for opaque
types to fail to take `@usableFromInline` into account and that results in the
emitted symbol getting the wrong linkage during IRGen. The fix is to make the
effective access computation take this quirk of opaque types into account
directly, instead of relying on copying of attributes.

Resolves rdar://110544170