[5.9] IRGen: alloc_global and global_addr instructions need to agree on the storage
The ABI used for `some` typed globals is similar to existential values.
We reserve an "inline" 3 word buffer, query type metadata's size, and if
the size is bigger allocate storage on the heap.

The `global_addr` instruction needs to follow this logic: If the storage is
opaque we need to project to the underlying buffer.

Risk: Low. The change should only affect opaque types (`some` types).

Without the change any `some` typed globals bigger than 3 words might
clobber storage in memory immediately after it.

rdar://109636344