[move-only] Remove use after move in CopiedLoadBorrowEliminationVisitor.
Specifically, we were calling process on a visitor that was already moved. To
fix this I created a separate state structure that the visitor initializes and
moved the post-processing step onto the state data structure.

This seems to not be causing any problems today, but could in the future.

rdar://111060475
(cherry picked from commit b4092c457258fd98ee3d4cbcdb742cb7144dac11)