sourcekitd: adjust the definition of `SOURCEKIT_PUBLIC`
The API surface for sourcekitd is exported through libAPI and
libService.  Use the `sourcekitd_EXPORTS` macro (reflecting the CMake
default behaviour of `_EXPORTS` macros) to provide a proper definition
of the macro on Windows.  This is not related to MSVC but to PE/COFF
which uses this to define the ABI boundary for the symbol.

This is helpful in partially repairing the ARM64 build of sourcekitd.